Extract common part of documentation to a new section - this closes #319
