+++ vertx-pg-client/src/main/asciidoc/index.adoc
@@ -143,179 +143,13 @@ $ PGUSER=user \
 {@link examples.PgClientExamples#configureFromEnv(io.vertx.core.Vertx)}
 ----
 
-== Running queries
+include::queries.adoc[]
 
-When you don't need a transaction or run single queries, you can run queries directly on the pool; the pool
-will use one of its connection to run the query and return the result to you.
+include::connections.adoc[]
 
-Here is how to run simple queries:
+include::transactions.adoc[]
 
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries01(io.vertx.sqlclient.SqlClient)}
-----
-
-You can do the same with prepared queries.
-
-The SQL string can refer to parameters by position, using `$1`, `$2`, etc…​
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries02(io.vertx.sqlclient.SqlClient)}
-----
-
-Query methods provides an asynchronous {@link io.vertx.sqlclient.RowSet} instance that works for _SELECT_ queries
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries03(io.vertx.sqlclient.SqlClient)}
-----
-
-or _UPDATE_/_INSERT_ queries:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries04(io.vertx.sqlclient.SqlClient)}
-----
-
-The {@link io.vertx.sqlclient.Row} gives you access to your data by index
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries05(Row)}
-----
-
-or by name
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries06(Row)}
-----
-
-You can access a wide variety of of types
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries07(Row)}
-----
-
-You can execute prepared batch
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries08(io.vertx.sqlclient.SqlClient)}
-----
-
-You can cache prepared queries:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries09(io.vertx.core.Vertx, PgConnectOptions, PoolOptions)}
-----
-
-You can fetch generated keys with a 'RETURNING' clause in your query:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#queries10(io.vertx.sqlclient.SqlClient)}
-----
-
-== Using connections
-
-When you need to execute sequential queries (without a transaction), you can create a new connection
-or borrow one from the pool:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#usingConnections01(io.vertx.core.Vertx, io.vertx.pgclient.PgPool)}
-----
-
-Prepared queries can be created:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#usingConnections02(io.vertx.pgclient.PgConnection)}
-----
-
-NOTE: prepared query caching depends on the {@link io.vertx.pgclient.PgConnectOptions#setCachePreparedStatements(boolean)} and
-does not depend on whether you are creating prepared queries or use {@link io.vertx.sqlclient.SqlClient#preparedQuery(java.lang.String, io.vertx.core.Handler) direct prepared queries}
-
-{@link io.vertx.sqlclient.PreparedQuery} can perform efficient batching:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#usingConnections03(io.vertx.pgclient.PgConnection)}
-----
-
-
-== Using transactions
-
-=== Transactions with connections
-
-You can execute transaction using SQL `BEGIN`/`COMMIT`/`ROLLBACK`, if you do so you must use
-a {@link io.vertx.pgclient.PgConnection} and manage it yourself.
-
-Or you can use the transaction API of {@link io.vertx.pgclient.PgConnection}:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#transaction01(io.vertx.pgclient.PgPool)}
-----
-
-When PostgreSQL reports the current transaction is failed (e.g the infamous _current transaction is aborted, commands ignored until
-end of transaction block_), the transaction is rollbacked and the {@link io.vertx.sqlclient.Transaction#abortHandler(io.vertx.core.Handler)}
-is called:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#transaction02(io.vertx.sqlclient.Transaction)}
-----
-
-=== Simplified transaction API
-
-When you use a pool, you can start a transaction directly on the pool.
-
-It borrows a connection from the pool, begins the transaction and releases the connection to the pool when the transaction ends.
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#transaction03(io.vertx.pgclient.PgPool)}
-----
-
-NOTE: this code will not close the connection because it will always be released back to the pool when the transaction
-
-== Cursors and streaming
-
-By default prepared query execution fetches all rows, you can use a
-{@link io.vertx.sqlclient.Cursor} to control the amount of rows you want to read:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#usingCursors01(io.vertx.pgclient.PgConnection)}
-----
-
-PostreSQL destroys cursors at the end of a transaction, so the cursor API shall be used
-within a transaction, otherwise you will likely get the `34000` PostgreSQL error.
-
-Cursors shall be closed when they are released prematurely:
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#usingCursors02(io.vertx.sqlclient.Cursor)}
-----
-
-A stream API is also available for cursors, which can be more convenient, specially with the Rxified version.
-
-[source,$lang]
-----
-{@link examples.PgClientExamples#usingCursors03(io.vertx.pgclient.PgConnection)}
-----
-
-The stream read the rows by batch of `50` and stream them, when the rows have been passed to the handler,
-a new batch of `50` is read and so on.
-
-The stream can be resumed or paused, the loaded rows will remain in memory until they are delivered and the cursor
-will stop iterating.
+include::cursor.adoc[]
 
 == PostgreSQL type mapping
