--- vertx-sql-client/src/main/asciidoc/index.adoc
@@ -1,3 +1,0 @@
-= Reactive SQL Client
-
-Todo.
